requirements.txt
pyrogram
telegraph
tgcrypto
python-dotenv